examples/basic_langgraph_agent/requirements.txt
# basic_langgraph_agent — dependencies
# Pin minor versions for reproducibility.
# Run: pip install -r requirements.txt

langgraph>=1.1.0,<2.0.0
langchain>=1.2.0,<2.0.0
langchain-openai>=1.1.0,<2.0.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
python-dotenv>=1.0.0
